Move groups API definitions to its app
